requirements/base.txt
django-cms>=2.4,<=2.5


